requirements.txt
asyncpg
discord.py
python-dotenv